qemu: special error code in case of no job on cancel block job
Special error code helps gracefully handle race conditions on
blockjob cancelling. Consider for example pull backup. We want
to stop it and in the process we cancel blockjob for some of the
exported disks. But at the time this command reaches qemu blockjob
fail for some reason and cancel result and result of stopping
will be an error with quite unexpecte message - no such job (sort of).
Instead we can detect this race and treat as correct cancelling
and wait until fail event will arrive. Then we can report corect
error message with some details from qemu probably.

Users of domainBlockJobAbort can be affected as -2 is new possible
return code. Not sure if I should stick to the original behaviour in
this case.
Message-Id: <1473240283-316922-3-git-send-email-nshirokovskiy@virtuozzo.com>