virt-admin: Tweak command parsing logic so that aliases point to new commands
Change the logic in a way, so that VSH_CMD_FLAG_ALIAS behaves similarly to
how VSH_OT_ALIAS for command options, i.e. there is no need for code duplication
for the alias and the aliased command structures. Along with that change,
switch any existing VSH_CMD_FLAG_ALIAS occurrences to this new format.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <87b49fbdcb928e10f211596e31158d4d3e9b6c65.1473667957.git.eskultet@redhat.com>