virhook: Adding inotify support to virhook.
Libvirtd only support hooks when the daemon is started. Hooks cannot be
loaded when the daemon is already running. So, to load a hook you need to
restart the service everytime. Now, the inotify support enables the option
of create a hook and run it even if libvirtd was started.

Cc: Carlos Castilho <ccasti@br.ibm.com>
Signed-off-by: Julio Faracco <jcfaracco@gmail.com>
Message-Id: <1473726116-8913-1-git-send-email-jcfaracco@gmail.com>