virt-admin: Tweak command parsing logic so that aliases point to new commands
Change the logic in a way, so that VSH_CMD_FLAG_ALIAS behaves similarly to
how VSH_OT_ALIAS for command options, i.e. there is no need for code duplication
for the alias and the aliased command structures. Along with that change,
switch any existing VSH_CMD_FLAG_ALIAS occurrences to this new format. Also,
since this patch introduces a new command structure element, adjust the
virsh-self-test test to make sure we won't ever miss to specify the '.alias'
member for an aliased command because doing that would lead to a segfault, as
would be the case for any missing element of the command structure.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <ff0a330217097f32a2ac6b3d2b7dcd00c88379c9.1473779393.git.eskultet@redhat.com>