qemu: Save various defaults for shmem
When we change the device used for the shared memory, it should not
change the settings, so rather save them upfront then having problems
later.

The only thing we are not saving is the role for old ivshmem and that's
because we were not specifying it, which means role=auto and that is
really bad idea to be using (due to various things).  However, we don't
want to change the behaviour, so that's the reason for that.

Details for the defaults of the newer implementation can be found in
qemu's commit 5400c02b90bb:

  http://git.qemu.org/?p=qemu.git;a=commit;h=5400c02b90bb

Signed-off-by: Martin Kletzander <mkletzan@redhat.com>
Message-Id: <6db3b08b25d640684680f68e8a1035e2dd4640f4.1473955636.git.mkletzan@redhat.com>