vsh: vshCmddefHelp: Retrieve command info after we know the command is non-NULL
Recent changes discarded an unnecessary 'else' block from vshCmddefHelp.
However, that also moved some var declarations as well as definitions to the
beginning of the function block and now we're trying to retrieve command's
.info element before actually checking whether we got a valid command or not.
So this patch fixes it accordingly. Also it removes variable 'help', since
there isn't any functional need for it.

Signed-off-by: Erik Skultety <eskultet@redhat.com>
Message-Id: <9d09e4935a5831f54eaa65382e5ff82af9911a57.1474021393.git.eskultet@redhat.com>