qemu: new functions qemuDomainDeviceConnectFlags*()
The lowest level function of this trio aims to be the ultimate
authority for the virDomainPCIConnectFlags to use for any given device
using a particular arch/machinetype/qemu-binary.

qemuDomainDeviceConnectFlagsInternal() returns the flags

qemuDomainDeviceConnectFlags() sets info->pciConnectFlags in a single
device (unless it has no virDomainDeviceInfo, in which case it's a
NOP).

qemuDomainDeviceSetAllConnectFlags() sets info->pciConnectFlags in all
devices that have a virDomainDeviceInfo (using
virDomainDeviceInfoIterate())

The latter two functions aren't called anywhere yet. This commit is
just making them available.
Message-Id: <1474398864-9039-5-git-send-email-laine@laine.org>