qemu: Split out filling of JSONGetBlockInfo data
Create a helper function which will read the device data and fill in a
hash table entry with that data.  In preparation for some more code reuse.

The helper function will take as an argument a new local static structure
in order to define the arguments. This will make it possible to combine
more code in later patches.

NB: Adjust error reporting to not use 'removable' or 'locked' as an
    argument to reporting error - rather just inline that in the message.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1475240034-11636-3-git-send-email-jferlan@redhat.com>