qemu: Combine the various ways to call query-block
Now that each of the 3 callers to "query-block" has relatively the same
flow, combine the remaining flow into qemuMonitorJSONQueryBlock and of
course rework the function comments to describe the expectations.

The qemuMonitorJSONQueryBlock will now take multiple parameters, "blindly"
building up the qemuMonitorJSONQueryBlockArgs, and then calling the
specific callback for each of the three consumers.

Each of the callers is essentially reduced to a shim calling the common
function with its specific qemuMonitorJSONQueryBlockFill* helper to
handle the needs for the specific caller.

NB: Also adjust the first "error message in the devices parsing loop
    to make it easier to distinguish which failure occurred.

Signed-off-by: John Ferlan <jferlan@redhat.com>
Message-Id: <1475240034-11636-6-git-send-email-jferlan@redhat.com>