conf: add support for choosing firmware type
Currently if you want to enable UEFI firmware for a guest
you need to know about the hypervisor platform specific
firmware path. This does not even work for all platforms,
as hypervisors like VMWare don't expose UEFI as a path,
they just have a direct config option for turning it on
or off.

This adds ability to use the much simpler:

  <loader firmware="uefi|bios"/>

to choose the different firmware types.

Signed-off-by: Daniel P. Berrange <berrange@redhat.com>
Message-Id: <1475509789-18443-3-git-send-email-berrange@redhat.com>